Fix search alias endpoints
