--- src/ApiPlatform/Resources/SearchAlias/DeleteSearchAlias.php
@@ -1,60 +1,0 @@
-<?php
-
-/**
- * Copyright since 2007 PrestaShop SA and Contributors
- * PrestaShop is an International Registered Trademark & Property of PrestaShop SA
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Academic Free License version 3.0
- * that is bundled with this package in the file LICENSE.md.
- * It is also available through the world-wide-web at this URL:
- * https://opensource.org/licenses/AFL-3.0
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@prestashop.com so we can send you a copy immediately.
- *
- * @author    PrestaShop SA and Contributors <contact@prestashop.com>
- * @copyright Since 2007 PrestaShop SA and Contributors
- * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
- */
-
-declare(strict_types=1);
-
-namespace PrestaShop\Module\APIResources\ApiPlatform\Resources\SearchAlias;
-
-use ApiPlatform\Metadata\ApiProperty;
-use ApiPlatform\Metadata\ApiResource;
-use PrestaShop\PrestaShop\Core\Domain\Alias\Command\DeleteSearchTermAliasesCommand;
-use PrestaShop\PrestaShop\Core\Domain\Alias\Exception\AliasConstraintException;
-use PrestaShop\PrestaShop\Core\Domain\Alias\Exception\AliasNotFoundException;
-use PrestaShopBundle\ApiPlatform\Metadata\CQRSDelete;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Validator\Constraints as Assert;
-
-#[ApiResource(
-    operations: [
-        new CQRSDelete(
-            // Usually the identifier would be in the URL, but since here it is a string this would conflict
-            // with the bulk url /search-aliases/bulk-delete (bulk-delete would be considered the identifier,
-            // vice versa) So the searchTerm will need to be passed in the JSON body
-            uriTemplate: '/search-aliases',
-            requirements: ['searchTerm' => '\d+'],
-            CQRSCommand: DeleteSearchTermAliasesCommand::class,
-            scopes: ['search_alias_write'],
-            allowEmptyBody: false,
-            experimentalOperation: true,
-        ),
-    ],
-    exceptionToStatus: [
-        AliasNotFoundException::class => Response::HTTP_NOT_FOUND,
-        AliasConstraintException::class => Response::HTTP_UNPROCESSABLE_ENTITY,
-    ],
-)]
-class DeleteSearchAlias
-{
-    #[ApiProperty(identifier: true)]
-    #[Assert\NotBlank]
-    #[Assert\Length(min: 1, max: 255)]
-    public string $searchTerm;
-}
